+++ tests/util_spec.rs
@@ -2,7 +2,7 @@ use std::fmt::Debug;
 
 use rstest::rstest;
 
-use beancount_parser_2::{parse, BeancountFile, DirectiveContent, Flag};
+use beancount_parser_2::{parse, BeancountFile, DirectiveContent};
 
 fn is_normal<T: Sized + Send + Sync + Unpin>() {}
 fn is_debug<T: Debug>() {}
@@ -13,19 +13,6 @@ fn beancount_file_type_should_be_normal() {
     is_normal::<BeancountFile<f32>>();
     is_debug::<BeancountFile<f32>>();
     is_clone::<BeancountFile<f32>>();
-}
-
-#[test]
-fn default_flag_is_completed() {
-    assert_eq!(Flag::default(), Flag::Completed);
-}
-
-#[rstest]
-#[case(Flag::Completed, '*')]
-#[case(Flag::Incomplete, '!')]
-fn can_convert_from_flag_to_char(#[case] flag: Flag, #[case] expected: char) {
-    let actual: char = flag.into();
-    assert_eq!(actual, expected);
 }
 
 #[test]